+++ src/components/Tabs/types.ts
@@ -1,0 +1,8 @@
+type TabConfig = {
+  label: string;
+  content: React.LazyExoticComponent<() => JSX.Element>;
+};
+
+export interface TabsProps {
+  tabs: TabConfig[];
+}
